Changed the Token api call into its own Apihandler class
- Changed the token api in to python class to reuse the get token method on both the Catalog api
	and the process api
- importet it in the main.py
